requirement.txt
Django==1.2.4
akismet==0.2.0
django-pagination==1.0.7
django-recaptcha-works==0.3.4
gunicorn==0.12.0
psycopg2==2.3.2
recaptcha-client==1.0.5
